% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_session.R
\name{browser_session_active}
\alias{browser_session_active}
\title{Check if a persistent browser session is active}
\usage{
browser_session_active()
}
\value{
Logical, TRUE if a session is active.
}
\description{
Check if a persistent browser session is active
}
\examples{
\dontrun{
browser_session_active()  # FALSE
browser_session_start()
browser_session_active()  # TRUE
browser_session_close()
browser_session_active()  # FALSE
}
}
