% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{check_hash}
\alias{check_hash}
\title{Check hash of a media file}
\usage{
check_hash(.x, hash_table, mediadir)
}
\arguments{
\item{.x}{Path to the media file.}

\item{hash_table}{A data frame of existing hashes.}

\item{mediadir}{Directory to save media files.}
}
\value{
A tibble with file details.
}
\description{
Check hash of a media file
}
