% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{create_necessary_dirs}
\alias{create_necessary_dirs}
\title{Create necessary directories}
\usage{
create_necessary_dirs(x)
}
\arguments{
\item{x}{Directory path to check and create.}
}
\value{
None (used for side effects).
}
\description{
Create necessary directories
}
