% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{detectmysnap}
\alias{detectmysnap}
\title{Updated function to detect the JSON code on Facebook ad websites}
\usage{
detectmysnap(rawhtmlascharacter)
}
\arguments{
\item{rawhtmlascharacter}{Raw HTML content as character string.}
}
\value{
A parsed JSON object.
}
\description{
Updated function to detect the JSON code on Facebook ad websites
}
