% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{detectmysnap_dep}
\alias{detectmysnap_dep}
\title{Detect the JSON code on Facebook ad websites}
\usage{
detectmysnap_dep(rawhtmlascharacter)
}
\arguments{
\item{rawhtmlascharacter}{Raw HTML content as character string.}
}
\value{
A parsed JSON object.
}
\description{
Function to detect the JSON code on facebook ad websites that contains the media URLs
This is basically str_extract but with perl!
}
\seealso{
\link{detectmysnap}
}
