% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ad_library_page_id.R
\name{.build_ad_library_page_url}
\alias{.build_ad_library_page_url}
\title{Build Ad Library URL from page ID (internal)}
\usage{
.build_ad_library_page_url(page_id, country = "ALL")
}
\arguments{
\item{page_id}{Character page ID.}

\item{country}{Character country code.}
}
\value{
Character scalar URL.
}
\description{
Build Ad Library URL from page ID (internal)
}
\keyword{internal}
