% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ad_library_page_id.R
\name{.extract_profile_transparency_data}
\alias{.extract_profile_transparency_data}
\title{Extract profile transparency data from active browser session (internal)}
\usage{
.extract_profile_transparency_data(b)
}
\arguments{
\item{b}{A chromote session object.}
}
\value{
List with \code{page_id} and \code{is_running_ads}.
}
\description{
Extract profile transparency data from active browser session (internal)
}
\keyword{internal}
