% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{download_media}
\alias{download_media}
\title{Download media files and potentially hash them}
\usage{
download_media(media_dat, mediadir = "data/media", hashing = T)
}
\arguments{
\item{media_dat}{Data containing media URLs.}

\item{mediadir}{Directory to save media files.}

\item{hashing}{Logical, whether to hash the files.}
}
\value{
None (used for side effects).
}
\description{
Download media files and potentially hash them
}
