% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{download_media_int}
\alias{download_media_int}
\title{Download media files with specified IDs}
\usage{
download_media_int(id, x, n, mediadir = "data/media")
}
\arguments{
\item{id}{Ad ID.}

\item{x}{Media URLs to download.}

\item{n}{Number of URLs.}

\item{mediadir}{Directory to save media files.}
}
\value{
A character vector with file paths.
}
\description{
Download media files with specified IDs
}
