% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{extract_media_urls}
\alias{extract_media_urls}
\title{Extract media URLs from data}
\usage{
extract_media_urls(yo)
}
\arguments{
\item{yo}{Data containing potential media URLs.}
}
\value{
A character vector of media URLs.
}
\description{
Extract media URLs from data
}
