% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{find_name}
\alias{find_name}
\title{Find an object in a nested list by name}
\usage{
find_name(haystack, needle)
}
\arguments{
\item{haystack}{A nested list.}

\item{needle}{Name of the object to find.}
}
\value{
Object in the nested list with the given name or NULL if not found.
}
\description{
Find an object in a nested list by name
}
