% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{get_ad_snapshots}
\alias{get_ad_snapshots}
\title{Get ad snapshots from Facebook ad library}
\usage{
get_ad_snapshots(ad_id, download = F, mediadir = "data/media", hashing = F)
}
\arguments{
\item{ad_id}{A character string specifying the ad ID.}

\item{download}{Logical, whether to download media files.}

\item{mediadir}{Directory to save media files.}

\item{hashing}{Logical, whether to hash the files. RECOMMENDED!}
}
\value{
A tibble with ad details.
}
\description{
Get ad snapshots from Facebook ad library
}
