% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ads_info.R
\name{get_ads_info}
\alias{get_ads_info}
\title{Get and Parse Ad Library Data}
\usage{
get_ads_info(ad_ids, country, keep_html = TRUE, cache_dir = "html_cache", ...)
}
\arguments{
\item{ad_ids}{A character vector of Ad Library IDs.}

\item{country}{A two-letter country code.}

\item{keep_html}{A logical flag. If \code{FALSE} (the default), the cache
directory with the downloaded HTML files will be deleted after parsing.
If \code{TRUE}, the files will be kept.}

\item{cache_dir}{The directory to store downloaded HTML files. Defaults to
"html_cache".}

\item{...}{Additional arguments to be passed down to \code{get_ad_html()}
(e.g., \code{overwrite}, \code{quiet}, \code{max_active}).}
}
\value{
A single, reordered dataframe containing the parsed ad data.
}
\description{
A wrapper function that downloads ad HTMLs for a given set of IDs and
a country, parses the data, and returns a final, reordered dataframe.
}
