% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_insights.R
\name{get_page_insights}
\alias{get_page_insights}
\title{Get Page Insights}
\usage{
get_page_insights(
  pageid,
  timeframe = "LAST_30_DAYS",
  lang = "en-GB",
  iso2c = "US",
  include_info = c("page_info", "targeting_info"),
  join_info = T
)
}
\arguments{
\item{pageid}{A string specifying the unique identifier of the Facebook page.}

\item{timeframe}{A string indicating the timeframe for the insights. Valid options include predefined
timeframes such as "LAST_30_DAYS". The default value is "LAST_30_DAYS".}

\item{lang}{A string representing the language locale to use for the request, formatted as language
code followed by country code (e.g., "en-GB" for English, United Kingdom). The default is "en-GB".}

\item{iso2c}{A string specifying the ISO-3166-1 alpha-2 country code for which insights are requested.
The default is "US".}

\item{include_info}{A character vector specifying the types of information to include in the output.
Possible values are "page_info" and "targeting_info". By default, both types of information are included.}

\item{join_info}{A logical value indicating whether to join page info and targeting info into a single
data frame (if TRUE) or return them as separate elements in a list (if FALSE). The default is TRUE.}
}
\value{
If \code{join_info} is TRUE, returns a data frame combining page and targeting information for
the specified Facebook page. If \code{join_info} is FALSE, returns a list with two elements:
\code{page_info} and \code{targeting_info}, each containing the respective data as a data frame.
In case of errors or no data available, the function may return a simplified data frame or list
indicating the absence of data.
}
\description{
Retrieves insights for a given Facebook page within a specified timeframe, language, and country.
It allows for fetching specific types of information and optionally joining page info with targeting info.
}
\examples{
insights <- get_page_insights(pageid="123456789", timeframe="LAST_30_DAYS", lang="en-GB", iso2c="US", 
                              include_info=c("page_info", "targeting_info"), join_info=TRUE)

}
