% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targeting.R
\name{get_targeting}
\alias{get_targeting}
\title{Get Meta Ad Library targeting data for a page}
\usage{
get_targeting(id, timeframe = "LAST_30_DAYS")
}
\arguments{
\item{id}{A character string representing the Facebook page ID.}

\item{timeframe}{A character string representing the desired timeframe. Can either be "LAST_30_DAYS" or "LAST_7_DAYS". Defaults to "LAST_30_DAYS".}
}
\value{
A \code{tibble} containing the audience data for the specified Facebook page and timeframe.
}
\description{
This function retrieves data for the targeting criteria of a Facebook page for a specified timeframe.
}
\examples{
\dontrun{
get_targeting("123456789")
get_targeting("987654321", "LAST_7_DAYS")
}

}
