% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{map_dfr_progress}
\alias{map_dfr_progress}
\title{Apply a function to each element of a list with a progress bar}
\usage{
map_dfr_progress(.x, .f, ...)
}
\arguments{
\item{.x}{List to iterate over.}

\item{.f}{Function to apply.}

\item{...}{Other parameters passed to \code{purrr::map_dfr}.}

\item{.id}{An identifier.}
}
\value{
An aggregated data frame.
}
\description{
Adding a progress bar to the map_dfr function https://www.jamesatkins.net/posts/progress-bar-in-purrr-map-df/
}
