% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_location.R
\name{parse_location}
\alias{parse_location}
\title{Parse Location from Ad Targeting Dataset}
\usage{
parse_location(.x, loc_var, type = "include", verbose = T)
}
\arguments{
\item{.x}{A data.frame containing the location string}

\item{loc_var}{A character string specifying the name of the column in .x that contains the location string}

\item{type}{A character string specifying the prefix to add to each column of split location details. Default is "include". Should be "include" or "exclude".}

\item{verbose}{A logical flag specifying whether to display a progress bar during processing. Default is \code{TRUE}.}
}
\value{
A data.frame with columns for each level of detail in the location.
}
\description{
A function to parse the location strings in the Ad Targeting Dataset and split into separate columns for each level of detail.
}
\examples{
\dontrun{
  ### create a dataset with unique include_location values
  distinct_data <- targeting_data \%>\%
    distinct(include_location)

  #### parse the location data and join in original dataset
  distinct_data \%>\%
    parse_location(include_location, type = "include") \%>\%
    right_join(targeting_data)

  ###----####

  ### create a dataset with unique exclude_location values
  distinct_data <- targeting_data \%>\%
    distinct(exclude_location)

  #### parse the location data and join in original dataset
  distinct_data \%>\%
    parse_location(exclude_location, type = "exclude") \%>\%
    right_join(targeting_data)

}
}
