% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_metatargetr_options}
\alias{set_metatargetr_options}
\title{Interactively Set and Save User Settings}
\usage{
set_metatargetr_options()
}
\description{
Launches an interactive command-line interface to help users configure
and save default package options, such as the cache directory and
user-agent randomization preference.
}
\details{
The function will guide the user through setting the following options:
\itemize{
\item \code{metatargetr.cache_dir}: The default directory to save HTML files.
\item \code{metatargetr.randomize_ua}: Whether to use random User-Agents by default.
}
The user will also be prompted to save these settings as environment variables
in their personal \code{.Renviron} file for persistence across R sessions.
}
