% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ad_report.R
\name{setup_playwright}
\alias{setup_playwright}
\title{Set up and initialize the playwrightr environment}
\usage{
setup_playwright()
}
\description{
Set up and initialize the playwrightr environment
}
\keyword{internal}
