% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{stupid_conversion}
\alias{stupid_conversion}
\title{Convert an entry for an ad into a row in a tibble}
\usage{
stupid_conversion(x)
}
\arguments{
\item{x}{An object containing data about the ad.}
}
\value{
A tibble row.
}
\description{
We did not manage to easily output the entry for one ad as a row in a tibble; this line solves the issue and does exactly that.
}
