% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_interests.R
\name{unnest_and_fix_dups}
\alias{unnest_and_fix_dups}
\title{Unnest interest targeting data}
\usage{
unnest_and_fix_dups(dat, the_list, new_name)
}
\arguments{
\item{dat}{a data frame}

\item{the_list}{the column name to unnest}

\item{new_name}{the name of the new column after unnesting and fixing duplicates}
}
\value{
a modified data frame with unnested and deduplicated values
}
\description{
unnest and fix duplicates in interest targeting data from Ad Targeting Dataset

The function unnests "include" and "exclude" columns in the Ad Targeting Dataset, and removes duplicates.
}
\examples{

### example usage:
## make sure you have the variable 'archive_id' in your data
ad_targeting_data \%>\%
    rowwise() \%>\%
    mutate(include_list = fix_json(include)) \%>\%
    ungroup() \%>\%
    ## the_list: the parsed list of JSON, new_name: what the parsed column should be called
   unnest_and_fix_dups(the_list = include_list, new_name = "parsed_include")
'
' @export
}
