% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media.R
\name{walk_progress}
\alias{walk_progress}
\title{Walk through a list with a progress bar}
\usage{
walk_progress(.x, .f, ...)
}
\arguments{
\item{.x}{List to iterate over.}

\item{.f}{Function to apply.}

\item{...}{Other parameters passed to \code{purrr::map_dfr}.}
}
\value{
None (used for side effects).
}
\description{
Walk through a list with a progress bar
}
